buildserver-notification: Use a foolproof indexing check
Previously the unstable repo had to be enabled and the use of sudo
was required to check if a package got indexed into the repo.

Now we download the index directly and compare to it if the build
succeeds. Additionally, this means we can always check when a package
gets indexed.