Makefile: Add safety catch for system.{base,devel} pkgs with used_libs additions
Previously it was very easy to accidently publish a system.base or
system.devel pkg which started to depend on another package. If
that package wasn't in system.{base,devel} the solbuild image would
fail to update, requiring manual intervention to rectify.

Now, if a pkg is part of system.{base,devel} and there are additions
to abi_used_libs the script will warn and prompt before publishing.

There are still improvements to be made here but it is a good starting
point to prevent repeated breakage.