rebuild-template-script: Improve utility of verify
Usage: ABI_LIB=libfoobar.so.7 ./foobar-rebuilds.sh verify

There are improvements to be made here, e.g. we can want to verify
changes in the pspec instead of abi_used_libs, or there may be several
possible libs it could have rebuilt against.

For now this is an improvement of the function.